==> code/2_descriptive stats.R <==
rm(list = ls())

library(tidyverse)
library(kableExtra)

### Descriptives on Zillow ----
load("data/zillow_county.RData")
source("code/1_data merging.R")

zillow_county_desc <- merge(select(aggregate(date ~ data_series, zillow_county, function(x){length(x)}), data_series, "observations" = date),
                            select(aggregate(date ~ data_series, zillow_county, min), data_series, "min_date" = date),
                            by = "data_series", all.x = TRUE)

zillow_county_desc <- merge(zillow_county_desc,
                            select(aggregate(date ~ data_series, zillow_county, max), data_series, "max_date" = date),
                            by = "data_series", all.x = TRUE)

zillow_county_desc$date_range <- paste0(zillow_county_desc$min_date, " - ", zillow_county_desc$max_date)
zillow_county_desc$min_date <- NULL
zillow_county_desc$max_date <- NULL

zillow_county$temp_fips <- fips_pad(zillow_county$state_code_fips, zillow_county$municipal_code_fips)
zillow_county_desc <- merge(zillow_county_desc,
                            select(aggregate(temp_fips ~ data_series, zillow_county, function(x){length(unique(x))}), data_series, "nr_regions" = temp_fips),
                            by = "data_series", all.x = TRUE)

zillow_county_desc <- merge(zillow_county_desc,
                            as.data.frame(as.matrix(aggregate(zhvi ~ data_series, zillow_county, function(x){quantile(x, c(0, 0.25, 0.5, 0.75, 1), na.rm = T)}))),
                            by = "data_series", all.x = TRUE)

zillow_county_desc <- merge(zillow_county_desc,
                            select(aggregate(zhvi ~ data_series, zillow_county, mean), data_series, "mean_zhvi" = zhvi),
                            by = "data_series", all.x = TRUE)

zillow_county_desc <- merge(zillow_county_desc,
                            select(aggregate(zhvi ~ data_series, zillow_county, function(x_){density(x_, na.rm = T)$x[which.max(density(x_, na.rm = T)$y)]}), data_series, "mode_zhvi" = zhvi),
                            by = "data_series", all.x = TRUE)

zillow_county_desc <- merge(zillow_county_desc,
                            select(aggregate(zhvi ~ data_series, zillow_county, var), data_series, "var_zhvi" = zhvi),
                            by = "data_series", all.x = TRUE)

zillow_county_desc <- merge(zillow_county_desc,
                            select(aggregate(zhvi ~ data_series, zillow_county, sd), data_series, "sd_zhvi" = zhvi),
                            by = "data_series", all.x = TRUE)

zillow_county_desc <- merge(zillow_county_desc,
                            select(aggregate(zhvi ~ data_series, zillow_county, moments::skewness), data_series, "skewness_zhvi" = zhvi),
                            by = "data_series", all.x = TRUE)

zillow_county_desc <- merge(zillow_county_desc,
                            select(aggregate(zhvi ~ data_series, zillow_county, moments::kurtosis), data_series, "kurtosis_zhvi" = zhvi),
                            by = "data_series", all.x = TRUE)

zillow_county_desc <- as.data.frame(t(zillow_county_desc))
names(zillow_county_desc) <- zillow_county_desc[1, ]
zillow_county_desc <- zillow_county_desc[-1,]
zillow_county_desc <- select(zillow_county_desc, all_homes_bottom_tier, all_homes_middle_tier, all_homes_top_tier, single_family_homes, condo_coop, one_bedroom, two_bedroom, three_bedroom, four_bedroom, five_plus_bedroom)
names(zillow_county_desc) <- c("Bottom Tier", "Middle Tier", "Top Tier", "Single Family", "Condo / Co-Op", "1 Bedroom", "2 Bedrooms", "3 Bedrooms", "4 Bedrooms", "5+ Bedrooms")
row.names(zillow_county_desc) <- c("Observations", "Date Range", "Regions", "Minimum", "25% Quantile", "Median", "75% Quantile", "Maximum", "Mean", "Mode", "Variance", "Standard Deviation", "Skewness", "Kurtosis")

# Save as Latex table
zillow_county_desc |> 
  kbl(format = "latex", align = "r") |> 
  kable_classic(full_width = TRUE) |> 
  write_file(file = "tables/zillow_descriptives.tex")


### Stationarity tests on Zillow ----
zillow_county_stat <- data.frame()
for (ds in unique(zillow_county$data_series)){
  z_sub <- subset(zillow_county, data_series == ds & !is.na(zhvi) & fips_code %in% subset(aggregate(date ~ fips_code + data_series, subset(zillow_county, !is.na(zhvi)), function(x){length(unique(x))}), date == 290 & data_series == ds)$fips_code)
  ips_test <- purtest(zhvi ~ trend, data = z_sub, index = c("fips_code", "date"), test = "ips", lags = "SIC")
  zillow_county_stat <- rbind(zillow_county_stat,
                              data.frame("Wtbar" = ips_test$statistic$statistic,
                                         "p_value" = ips_test$statistic$p.value,
                                         "data_series" = ds,
                                         "comment" = "Only counties with 290 observations."))
  print(ds)
}

openxlsx::write.xlsx(zillow_county_stat, file = "tables/zillow_stationarity.xlsx")


### Descriptives on FEMA ----
rm(list = ls())
gc()

load("data/fema.RData")
load("data/zillow_county.RData")
source("code/1_data merging.R")

# fema <- merge(fema, fema_incident_mapping, by = "incident_type", all.x = TRUE)
fema <- merge(fema, fema_incident_mapping_final, by = "incident_type", all.x = TRUE)

fema_desc <- aggregate(incident_type ~ incident_category_man, fema, function(x){paste(unique(x), collapse = ", ")})
fema_desc$incident_category_man <- factor(fema_desc$incident_category_man, levels = c("Cold", "Fire", "Geological", "Human Cause", "Water", "Wind", "Other"))

fema_desc <- merge(fema_desc,
                   select(aggregate(disaster_number ~ incident_category_man, subset(fema, date_incident_begin >= as.Date("2000-01-01")), function(x){length(unique(x))}), incident_category_man, "nr_disasters" = disaster_number),
                   by = "incident_category_man", all.x = TRUE)

fema_desc <- merge(fema_desc,
                   select(aggregate(date_declaration ~ incident_category_man, subset(fema, date_incident_begin >= as.Date("2000-01-01")), function(x){paste0(min(x), " - ", max(x))}), incident_category_man, "date_range" = date_declaration),
                   by = "incident_category_man", all.x = TRUE)

zillow_county$fips_code <- fips_pad(zillow_county$state_code_fips, zillow_county$municipal_code_fips)

fema_desc <- merge(fema_desc,
                   select(aggregate(place_code ~ incident_category_man, subset(fema, date_incident_begin >= as.Date("2000-01-01")), function(x){length(unique(x)[unique(x) %in% zillow_county$fips_code])}), incident_category_man, "regions" = place_code),
                   by = "incident_category_man", all.x = TRUE)

# Save as Latex table
fema_desc |> 
  kbl(format = "latex", align = "r") |> 
  kable_classic(full_width = TRUE) |> 
  write_file(file = "tables/fema_descriptives.tex")


load("data/spec_5_1_panel_maj_fg.RData")
spec_5_1_panel_maj <- pivot_longer(select(subset(spec_5_1_panel_maj, data_series == "all_homes_middle_tier"), fips_code, date, nr_dis_lag_0.25, nr_dis_lag_0.5, nr_dis_lag_0.5_e, nr_dis_lag_1_e), cols = -c(fips_code, date), names_to = "time_window", values_to = "nr_dis")

dis_desc <- merge(select(aggregate(date ~ time_window, spec_5_1_panel_maj, function(x){length(x)}), time_window, "observations" = date),
                  select(aggregate(date ~ time_window, spec_5_1_panel_maj, min), time_window, "min_date" = date),
                  by = "time_window", all.x = TRUE)

dis_desc <- merge(dis_desc,
                  select(aggregate(date ~ time_window, spec_5_1_panel_maj, max), time_window, "max_date" = date),
                  by = "time_window", all.x = TRUE)

dis_desc$date_range <- paste0(dis_desc$min_date, " - ", dis_desc$max_date)
dis_desc$min_date <- NULL
dis_desc$max_date <- NULL

dis_desc <- merge(dis_desc,
                  select(aggregate(fips_code ~ time_window, spec_5_1_panel_maj, function(x){length(unique(x))}), time_window, "nr_regions" = fips_code),
                  by = "time_window", all.x = TRUE)

dis_desc <- merge(dis_desc,
                  as.data.frame(as.matrix(aggregate(nr_dis ~ time_window, spec_5_1_panel_maj, function(x){quantile(x, c(0, 0.25, 0.5, 0.75, 1), na.rm = T)}))),
                  by = "time_window", all.x = TRUE)

dis_desc <- merge(dis_desc,
                  select(aggregate(nr_dis ~ time_window, spec_5_1_panel_maj, mean), time_window, "mean_nr_dis" = nr_dis),
                  by = "time_window", all.x = TRUE)

dis_desc <- merge(dis_desc,
                  select(aggregate(nr_dis ~ time_window, spec_5_1_panel_maj, function(x_){density(x_, na.rm = T)$x[which.max(density(x_, na.rm = T)$y)]}), time_window, "mode_nr_dis" = nr_dis),
                  by = "time_window", all.x = TRUE)

dis_desc <- merge(dis_desc,
                  select(aggregate(nr_dis ~ time_window, spec_5_1_panel_maj, var), time_window, "var_nr_dis" = nr_dis),
                  by = "time_window", all.x = TRUE)

dis_desc <- merge(dis_desc,
                  select(aggregate(nr_dis ~ time_window, spec_5_1_panel_maj, sd), time_window, "sd_nr_dis" = nr_dis),
                  by = "time_window", all.x = TRUE)

dis_desc <- merge(dis_desc,
                  select(aggregate(nr_dis ~ time_window, spec_5_1_panel_maj, moments::skewness), time_window, "skewness_nr_dis" = nr_dis),
                  by = "time_window", all.x = TRUE)

dis_desc <- merge(dis_desc,
                  select(aggregate(nr_dis ~ time_window, spec_5_1_panel_maj, moments::kurtosis), time_window, "kurtosis_nr_dis" = nr_dis),
                  by = "time_window", all.x = TRUE)

dis_desc <- as.data.frame(t(dis_desc))
names(dis_desc) <- dis_desc[1, ]
dis_desc <- dis_desc[-1,]
dis_desc <- select(dis_desc, nr_dis_lag_0.25, nr_dis_lag_0.5, nr_dis_lag_0.5_e, nr_dis_lag_1_e)
row.names(dis_desc) <- c("Observations", "Date Range", "Regions", "Minimum", "25% Quantile", "Median", "75% Quantile", "Maximum", "Mean", "Mode", "Variance", "Standard Deviation", "Skewness", "Kurtosis")

# Save as Latex table
dis_desc |> 
  kbl(format = "latex", align = "r") |> 
  kable_classic(full_width = TRUE) |> 
  write_file(file = "tables/nr_dis_descriptives.tex")


load("data/decl_type_panel.RData")
decl_type_panel <- select(decl_type_panel, fips_code, date, perc_maj_0.25, perc_maj_0.5, perc_maj_0.5_e, perc_maj_1_e)
decl_type_panel$perc_maj_0.25[is.na(decl_type_panel$perc_maj_0.25)] <- 0
decl_type_panel$perc_maj_0.5[is.na(decl_type_panel$perc_maj_0.5)] <- 0
decl_type_panel$perc_maj_0.5_e[is.na(decl_type_panel$perc_maj_0.5_e)] <- 0
decl_type_panel$perc_maj_1_e[is.na(decl_type_panel$perc_maj_1_e)] <- 0
decl_type_panel$perc_maj_0.25 <- decl_type_panel$perc_maj_0.25 * 100
decl_type_panel$perc_maj_0.5 <- decl_type_panel$perc_maj_0.5 * 100
decl_type_panel$perc_maj_0.5_e <- decl_type_panel$perc_maj_0.5_e * 100
decl_type_panel$perc_maj_1_e <- decl_type_panel$perc_maj_1_e * 100

decl_type_panel <- pivot_longer(decl_type_panel, cols = -c(fips_code, date), names_to = "time_window", values_to = "perc_maj")

maj_desc <- merge(select(aggregate(date ~ time_window, decl_type_panel, function(x){length(x)}), time_window, "observations" = date),
                  select(aggregate(date ~ time_window, decl_type_panel, min), time_window, "min_date" = date),
                  by = "time_window", all.x = TRUE)

maj_desc <- merge(maj_desc,
                  select(aggregate(date ~ time_window, decl_type_panel, max), time_window, "max_date" = date),
                  by = "time_window", all.x = TRUE)

maj_desc$date_range <- paste0(maj_desc$min_date, " - ", maj_desc$max_date)
maj_desc$min_date <- NULL
maj_desc$max_date <- NULL

maj_desc <- merge(maj_desc,
                  select(aggregate(fips_code ~ time_window, decl_type_panel, function(x){length(unique(x))}), time_window, "nr_regions" = fips_code),
                  by = "time_window", all.x = TRUE)

maj_desc <- merge(maj_desc,
                  as.data.frame(as.matrix(aggregate(perc_maj ~ time_window, decl_type_panel, function(x){quantile(x, c(0, 0.25, 0.5, 0.75, 1), na.rm = T)}))),
                  by = "time_window", all.x = TRUE)

maj_desc <- merge(maj_desc,
                  select(aggregate(perc_maj ~ time_window, decl_type_panel, mean), time_window, "mean_nr_dis" = perc_maj),
                  by = "time_window", all.x = TRUE)

maj_desc <- merge(maj_desc,
                  select(aggregate(perc_maj ~ time_window, decl_type_panel, function(x_){density(x_, na.rm = T)$x[which.max(density(x_, na.rm = T)$y)]}), time_window, "mode_nr_dis" = perc_maj),
                  by = "time_window", all.x = TRUE)

maj_desc <- merge(maj_desc,
                  select(aggregate(perc_maj ~ time_window, decl_type_panel, var), time_window, "var_nr_dis" = perc_maj),
                  by = "time_window", all.x = TRUE)

maj_desc <- merge(maj_desc,
                  select(aggregate(perc_maj ~ time_window, decl_type_panel, sd), time_window, "sd_nr_dis" = perc_maj),
                  by = "time_window", all.x = TRUE)

maj_desc <- merge(maj_desc,
                  select(aggregate(perc_maj ~ time_window, decl_type_panel, moments::skewness), time_window, "skewness_nr_dis" = perc_maj),
                  by = "time_window", all.x = TRUE)

maj_desc <- merge(maj_desc,
                  select(aggregate(perc_maj ~ time_window, decl_type_panel, moments::kurtosis), time_window, "kurtosis_nr_dis" = perc_maj),
                  by = "time_window", all.x = TRUE)

maj_desc <- as.data.frame(t(maj_desc))
names(maj_desc) <- maj_desc[1, ]
maj_desc <- maj_desc[-1,]
maj_desc <- select(maj_desc, perc_maj_0.25, perc_maj_0.5, perc_maj_0.5_e, perc_maj_1_e)
row.names(maj_desc) <- c("Observations", "Date Range", "Regions", "Minimum", "25% Quantile", "Median", "75% Quantile", "Maximum", "Mean", "Mode", "Variance", "Standard Deviation", "Skewness", "Kurtosis")

# Save as Latex table
maj_desc |> 
  kbl(format = "latex", align = "r") |> 
  kable_classic(full_width = TRUE) |> 
  write_file(file = "tables/perc_maj_descriptives.tex")


### Descriptives on BEA / LAUS ----
load("data/bea_gdp.RData")
load("data/bls_laus.RData")
load("data/bls_qcew.RData")

laus$fips_code <- fips_pad(laus$state_code, laus$county_code)

econ_desc <- merge(merge(subset(select(qcew, "fips_code" = area_fips, agglvl_code, year, qtr, avg_wkly_wage), agglvl_code == 70 & avg_wkly_wage != 0, select = c(fips_code, year, qtr, avg_wkly_wage)),
                         select(laus, fips_code, year, unemployment_rate), by = c("fips_code", "year"), all.x = TRUE),
                   subset(bea, table_name == "CAGDP2" & industry_classification == "...", select = c(fips_code, gdp_value, year)), by = c("fips_code", "year"), all.x = TRUE) |> 
  subset(year >= "2000") |> 
  mutate("unemployment_rate" = as.numeric(unemployment_rate), # NAs introduced through 368 "N.A." strings, checked and okay
         "gdp_value" = as.numeric(gdp_value)) |> 
  pivot_longer(cols = c("avg_wkly_wage", "unemployment_rate", "gdp_value"), names_to = "data_series", values_to = "econ_value") |> 
  mutate("date" = paste0(year, "-", str_pad(qtr, width = 2, side = "left", pad = "0")))

econ_desc_out <- merge(select(aggregate(econ_value ~ data_series, econ_desc, function(x){length(na.omit(x))}), data_series, "observations" = econ_value),
                       select(aggregate(date ~ data_series, subset(econ_desc, !is.na(econ_value)), min), data_series, "min_date" = date),
                       by = "data_series", all.x = TRUE)

econ_desc_out <- merge(econ_desc_out,
                       select(aggregate(date ~ data_series, subset(econ_desc, !is.na(econ_value)), max), data_series, "max_date" = date),
                       by = "data_series", all.x = TRUE)

econ_desc_out$date_range <- paste0(econ_desc_out$min_date, " - ", econ_desc_out$max_date)
econ_desc_out$min_date <- NULL
econ_desc_out$max_date <- NULL

econ_desc_out <- merge(econ_desc_out,
                       select(aggregate(fips_code ~ data_series, subset(econ_desc, !is.na(econ_value)), function(x){length(unique(x))}), data_series, "nr_regions" = fips_code),
                       by = "data_series", all.x = TRUE)

econ_desc_out <- merge(econ_desc_out,
                       as.data.frame(as.matrix(aggregate(econ_value ~ data_series, subset(econ_desc, !is.na(econ_value)), function(x){quantile(x, c(0, 0.25, 0.5, 0.75, 1), na.rm = T)}))),
                       by = "data_series", all.x = TRUE)

econ_desc_out <- merge(econ_desc_out,
                       select(aggregate(econ_value ~ data_series, subset(econ_desc, !is.na(econ_value)), mean), data_series, "mean_econ_value" = econ_value),
                       by = "data_series", all.x = TRUE)

econ_desc_out <- merge(econ_desc_out,
                       select(aggregate(econ_value ~ data_series, subset(econ_desc, !is.na(econ_value)), function(x_){density(x_, na.rm = T)$x[which.max(density(x_, na.rm = T)$y)]}), data_series, "mode_econ_value" = econ_value),
                       by = "data_series", all.x = TRUE)

econ_desc_out <- merge(econ_desc_out,
                       select(aggregate(econ_value ~ data_series, subset(econ_desc, !is.na(econ_value)), var), data_series, "var_econ_value" = econ_value),
                       by = "data_series", all.x = TRUE)

econ_desc_out <- merge(econ_desc_out,
                       select(aggregate(econ_value ~ data_series, subset(econ_desc, !is.na(econ_value)), sd), data_series, "sd_econ_value" = econ_value),
                       by = "data_series", all.x = TRUE)

econ_desc_out <- merge(econ_desc_out,
                       select(aggregate(econ_value ~ data_series, subset(econ_desc, !is.na(econ_value)), moments::skewness), data_series, "skewness_econ_value" = econ_value),
                       by = "data_series", all.x = TRUE)

econ_desc_out <- merge(econ_desc_out,
                       select(aggregate(econ_value ~ data_series, subset(econ_desc, !is.na(econ_value)), moments::kurtosis), data_series, "kurtosis_econ_value" = econ_value),
                       by = "data_series", all.x = TRUE)

econ_desc_out <- as.data.frame(t(econ_desc_out))
names(econ_desc_out) <- econ_desc_out[1, ]
econ_desc_out <- econ_desc_out[-1,]
econ_desc_out <- select(econ_desc_out, unemployment_rate, avg_wkly_wage, gdp_value)
names(econ_desc_out) <- c("Unemployment rate", "Average weekly wage", "GDP")
row.names(econ_desc_out) <- c("Observations", "Date Range", "Regions", "Minimum", "25% Quantile", "Median", "75% Quantile", "Maximum", "Mean", "Mode", "Variance", "Standard Deviation", "Skewness", "Kurtosis")

# Save as Latex table
econ_desc_out |> 
  kbl(format = "latex", align = "r") |> 
  kable_classic(full_width = TRUE) |> 
  write_file(file = "tables/econ_descriptives.tex")



### Descriptives on HMDA ----
load("data/hmda_panels.RData")

hmda <- merge(dti_panel_median, ltv_panel_median, by = c("fips_code", "as_of_year"), all = T)
hmda$median_debt_to_income_ratio <- as.numeric(hmda$median_debt_to_income_ratio)
hmda$median_loan_to_value_ratio <- as.numeric(hmda$median_loan_to_value_ratio)

hmda <- pivot_longer(hmda, cols = c("median_debt_to_income_ratio", "median_loan_to_value_ratio"), names_to = "data_series", values_to = "value")
hmda <- subset(hmda, value > 0)

hmda_desc <- merge(select(aggregate(as_of_year ~ data_series, hmda, function(x){length(na.omit(x))}), data_series, "observations" = as_of_year),
                   select(aggregate(as_of_year ~ data_series, hmda, min), data_series, "min_date" = as_of_year),
                   by = "data_series", all.x = TRUE)

hmda_desc <- merge(hmda_desc,
                   select(aggregate(as_of_year ~ data_series, hmda, max), data_series, "max_date" = as_of_year),
                   by = "data_series", all.x = TRUE)

hmda_desc$date_range <- paste0(hmda_desc$min_date, " - ", hmda_desc$max_date)
hmda_desc$min_date <- NULL
hmda_desc$max_date <- NULL

hmda_desc <- merge(hmda_desc,
                   select(aggregate(fips_code ~ data_series, hmda, function(x){length(unique(x))}), data_series, "nr_regions" = fips_code),
                   by = "data_series", all.x = TRUE)

hmda_desc <- merge(hmda_desc,
                   as.data.frame(as.matrix(aggregate(value ~ data_series, hmda, function(x){quantile(x, c(0, 0.25, 0.5, 0.75, 1), na.rm = T)}))),
                   by = "data_series", all.x = TRUE)

hmda_desc <- merge(hmda_desc,
                   select(aggregate(value ~ data_series, hmda, mean), data_series, "mean_value" = value),
                   by = "data_series", all.x = TRUE)

hmda_desc <- merge(hmda_desc,
                   select(aggregate(value ~ data_series, hmda, function(x_){density(x_, na.rm = T)$x[which.max(density(x_, na.rm = T)$y)]}), data_series, "mode_value" = value),
                   by = "data_series", all.x = TRUE)

hmda_desc <- merge(hmda_desc,
                   select(aggregate(value ~ data_series, hmda, var), data_series, "var_value" = value),
                   by = "data_series", all.x = TRUE)

hmda_desc <- merge(hmda_desc,
                   select(aggregate(value ~ data_series, hmda, sd), data_series, "sd_value" = value),
                   by = "data_series", all.x = TRUE)

hmda_desc <- merge(hmda_desc,
                   select(aggregate(value ~ data_series, hmda, moments::skewness), data_series, "skewness_value" = value),
                   by = "data_series", all.x = TRUE)

hmda_desc <- merge(hmda_desc,
                   select(aggregate(value ~ data_series, hmda, moments::kurtosis), data_series, "kurtosis_value" = value),
                   by = "data_series", all.x = TRUE)

hmda_desc <- as.data.frame(t(hmda_desc))
names(hmda_desc) <- hmda_desc[1, ]
hmda_desc <- hmda_desc[-1,]
names(hmda_desc) <- c("Median DTI", "Median LTV")
row.names(hmda_desc) <- c("Observations", "Date Range", "Regions", "Minimum", "25% Quantile", "Median", "75% Quantile", "Maximum", "Mean", "Mode", "Variance", "Standard Deviation", "Skewness", "Kurtosis")

# Save as Latex table
hmda_desc |> 
  kbl(format = "latex", align = "r") |> 
  kable_classic(full_width = TRUE) |> 
  write_file(file = "tables/hmda_descriptives.tex")



### Descriptives on NRI ----
load("data/nri.RData")

nri_composite$fips_code <- fips_pad(nri_composite$state_code, nri_composite$county_code)

nri <- select(nri_composite, fips_code, risk_score_composite, alr_building_composite)
nri$alr_building_composite <- nri$alr_building_composite * 100
nri <- pivot_longer(nri, cols = -fips_code, names_to = "data_series", values_to = "value")
nri <- unique(nri)

nri_desc <- merge(select(aggregate(value ~ data_series, nri, function(x){length(na.omit(x))}), data_series, "observations" = value),
                  as.data.frame(as.matrix(aggregate(value ~ data_series, nri, function(x){quantile(x, c(0, 0.25, 0.5, 0.75, 1), na.rm = T)}))),
                  by = "data_series", all.x = TRUE)

nri_desc <- merge(nri_desc,
                  select(aggregate(value ~ data_series, nri, mean), data_series, "mean_value" = value),
                  by = "data_series", all.x = TRUE)

nri_desc <- merge(nri_desc,
                  select(aggregate(value ~ data_series, nri, function(x_){density(x_, na.rm = T)$x[which.max(density(x_, na.rm = T)$y)]}), data_series, "mode_value" = value),
                  by = "data_series", all.x = TRUE)

nri_desc <- merge(nri_desc,
                  select(aggregate(value ~ data_series, nri, var), data_series, "var_value" = value),
                  by = "data_series", all.x = TRUE)

nri_desc <- merge(nri_desc,
                  select(aggregate(value ~ data_series, nri, sd), data_series, "sd_value" = value),
                  by = "data_series", all.x = TRUE)

nri_desc <- merge(nri_desc,
                  select(aggregate(value ~ data_series, nri, moments::skewness), data_series, "skewness_value" = value),
                  by = "data_series", all.x = TRUE)

nri_desc <- merge(nri_desc,
                  select(aggregate(value ~ data_series, nri, moments::kurtosis), data_series, "kurtosis_value" = value),
                  by = "data_series", all.x = TRUE)

nri_desc <- as.data.frame(t(nri_desc))
names(nri_desc) <- nri_desc[1, ]
nri_desc <- nri_desc[-1,]
names(nri_desc) <- c("ALR buildings", "Risk score")
row.names(nri_desc) <- c("Observations", "Minimum", "25% Quantile", "Median", "75% Quantile", "Maximum", "Mean", "Mode", "Variance", "Standard Deviation", "Skewness", "Kurtosis")

# Save as Latex table
nri_desc |> 
  kbl(format = "latex", align = "r") |> 
  kable_classic(full_width = TRUE) |> 
  write_file(file = "tables/nri_descriptives.tex")
