==> code/5.1_robustness.R <==
rm(list = ls())

library(tidyverse)
library(plm)

source("code/1_data merging.R")

# Definitions
incident_type_excl <- c("Biological", "Chemical", "Other", "Toxic Substances")
time_period_spec <- list("pre_gfc" = c("2000-01-31", "2007-12-31"),
                         "post_gfc" = c("2009-01-31", "2019-12-31"),
                         "post_covid" = c("2021-01-31", "2024-02-29"),
                         "bush" = c("2001-01-31", "2008-12-31"),
                         "obama" = c("2009-01-31", "2016-12-31"),
                         "trump" = c("2017-01-31", "2020-12-31"),
                         "biden" = c("2021-01-31", "2024-02-29"))

### Spec 6.1 (Econ time frames) ----
load("data/spec_5_3_panel_maj_wide.RData")

spec_6_1_final_maj <- rbind(data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & date >= time_period_spec$pre_gfc[1] & date <= time_period_spec$pre_gfc[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "time_period" = "pre_gfc", "spec" = "6.1"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & date >= time_period_spec$post_gfc[1] & date <= time_period_spec$post_gfc[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "time_period" = "post_gfc", "spec" = "6.1"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & date >= time_period_spec$post_covid[1] & date <= time_period_spec$post_covid[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "time_period" = "post_covid", "spec" = "6.1"))

openxlsx::write.xlsx(spec_6_1_final_maj, file = "results/final/spec_6_1_final_maj.xlsx")


### Spec 6.2 (Political time frames) ----
spec_6_2_final_maj <- rbind(data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & date >= time_period_spec$bush[1] & date <= time_period_spec$bush[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "time_period" = "bush", "spec" = "6.2"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & date >= time_period_spec$obama[1] & date <= time_period_spec$obama[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "time_period" = "obama", "spec" = "6.2"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & date >= time_period_spec$trump[1] & date <= time_period_spec$trump[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "time_period" = "trump", "spec" = "6.2"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & date >= time_period_spec$biden[1] & date <= time_period_spec$biden[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "time_period" = "biden", "spec" = "6.2"))

openxlsx::write.xlsx(spec_6_2_final_maj, file = "results/final/spec_6_2_final_maj.xlsx")


### Spec 6.3 (Credit constraint / credit quality using HMDA) ----
load("data/hmda_panels.RData")

ltv_panel_median$median_loan_to_value_ratio <- as.numeric(ltv_panel_median$median_loan_to_value_ratio)
dti_panel_median$median_debt_to_income_ratio <- as.numeric(dti_panel_median$median_debt_to_income_ratio) # NAs introduced from empty or range values, checked and okay

ltv_panel_median <- subset(ltv_panel_median, !is.na(median_loan_to_value_ratio) & median_loan_to_value_ratio > 0)
dti_panel_median <- subset(dti_panel_median, !is.na(median_debt_to_income_ratio) & median_loan_to_value_ratio > 0)

spec_5_3_panel_maj$year <- lubridate::year(spec_5_3_panel_maj$date)
spec_5_3_panel_maj <- merge(spec_5_3_panel_maj,
                            ltv_panel_median,
                            by.x = c("fips_code", "year"), by.y = c("fips_code", "as_of_year"), all.x = T)
spec_5_3_panel_maj <- merge(spec_5_3_panel_maj,
                            dti_panel_median,
                            by.x = c("fips_code", "year"), by.y = c("fips_code", "as_of_year"), all.x = T)

ltv_mapping <- list("ltv_low" = c(0, 80),
                    "ltv_medium" = c(80, 95),
                    "ltv_high" = c(95, 100))

dti_mapping <- list("dti_low" = c(0, 35),
                    "dti_medium" = c(35, 50),
                    "dti_high" = c(50, 999))

spec_6_3_final_maj <- rbind(data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & median_loan_to_value_ratio > ltv_mapping$ltv_low[1] & median_loan_to_value_ratio <= ltv_mapping$ltv_low[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "median_ltv_low", "spec" = "6.3"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & median_loan_to_value_ratio > ltv_mapping$ltv_medium[1] & median_loan_to_value_ratio <= ltv_mapping$ltv_medium[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "median_ltv_medium", "spec" = "6.3"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & median_loan_to_value_ratio > ltv_mapping$ltv_high[1] & median_loan_to_value_ratio <= ltv_mapping$ltv_high[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "median_ltv_high", "spec" = "6.3"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & median_debt_to_income_ratio > dti_mapping$dti_low[1] & median_debt_to_income_ratio <= dti_mapping$dti_low[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "median_dti_low", "spec" = "6.3"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & median_debt_to_income_ratio > dti_mapping$dti_medium[1] & median_debt_to_income_ratio <= dti_mapping$dti_medium[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "median_dti_medium", "spec" = "6.3"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & median_debt_to_income_ratio > dti_mapping$dti_high[1] & median_debt_to_income_ratio <= dti_mapping$dti_high[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "median_dti_high", "spec" = "6.3"))

openxlsx::write.xlsx(spec_6_3_final_maj, file = "results/final/spec_6_3_final_maj.xlsx")

# Compile non-amortizing loan features as subprime proxies
subprime_panel <- subset(subprime_panel, as_of_year >= 2018)

spec_5_3_panel_maj <- merge(spec_5_3_panel_maj,
                            subprime_panel,
                            by.x = c("fips_code", "year"), by.y = c("fips_code", "as_of_year"), all.x = T)

spec_5_3_panel_maj$perc_irp[is.na(spec_5_3_panel_maj$perc_irp)] <- 0
spec_5_3_panel_maj$perc_namo[is.na(spec_5_3_panel_maj$perc_namo)] <- 0
spec_5_3_panel_maj$perc_iop[is.na(spec_5_3_panel_maj$perc_iop)] <- 0
spec_5_3_panel_maj$perc_at_least_one_namo[is.na(spec_5_3_panel_maj$perc_at_least_one_namo)] <- 0
spec_5_3_panel_maj$perc_at_least_two_namo[is.na(spec_5_3_panel_maj$perc_at_least_two_namo)] <- 0
spec_5_3_panel_maj$perc_three_namo[is.na(spec_5_3_panel_maj$perc_three_namo)] <- 0

subprime_mapping <- list("irp_low" = c(0, median(subset(subprime_panel, perc_irp > 0)$perc_irp)),
                         "irp_high" = c(median(subset(subprime_panel, perc_irp > 0)$perc_irp), 1),
                         "iop_low" = c(0, median(subset(subprime_panel, perc_iop > 0)$perc_iop)),
                         "iop_high" = c(median(subset(subprime_panel, perc_iop > 0)$perc_iop), 1),
                         "namo_low" = c(0, median(subset(subprime_panel, perc_namo > 0)$perc_namo)),
                         "namo_high" = c(median(subset(subprime_panel, perc_namo > 0)$perc_namo), 1),
                         "sp_one_low" = c(0, median(subset(subprime_panel, perc_at_least_one_namo > 0)$perc_at_least_one_namo)),
                         "sp_one_high" = c(median(subset(subprime_panel, perc_at_least_one_namo > 0)$perc_at_least_one_namo), 1),
                         "sp_two_low" = c(0, median(subset(subprime_panel, perc_at_least_two_namo > 0)$perc_at_least_two_namo)),
                         "sp_two_high" = c(median(subset(subprime_panel, perc_at_least_two_namo > 0)$perc_at_least_two_namo), 1),
                         "sp_three_low" = c(0, median(subset(subprime_panel, perc_three_namo > 0)$perc_three_namo)),
                         "sp_three_high" = c(median(subset(subprime_panel, perc_three_namo > 0)$perc_three_namo), 1))

spec_6_3_extensions_maj <- rbind(data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_irp > subprime_mapping$irp_low[1] & perc_irp <= subprime_mapping$irp_low[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_irp_low", "spec" = "6.3_e"),
                                 data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_irp > subprime_mapping$irp_high[1] & perc_irp <= subprime_mapping$irp_high[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_irp_high", "spec" = "6.3_e"),
                                 data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_iop > subprime_mapping$iop_low[1] & perc_iop <= subprime_mapping$iop_low[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_iop_low", "spec" = "6.3_e"),
                                 data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_iop > subprime_mapping$iop_high[1] & perc_iop <= subprime_mapping$iop_high[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_iop_high", "spec" = "6.3_e"),
                                 data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_namo > subprime_mapping$namo_low[1] & perc_namo <= subprime_mapping$namo_low[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_namo_low", "spec" = "6.3_e"),
                                 data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_namo > subprime_mapping$namo_high[1] & perc_namo <= subprime_mapping$namo_high[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_namo_high", "spec" = "6.3_e"),
                                 data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_at_least_one_namo > subprime_mapping$sp_one_low[1] & perc_at_least_one_namo <= subprime_mapping$sp_one_low[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_at_least_one_low", "spec" = "6.3_e"),
                                 data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_at_least_one_namo > subprime_mapping$sp_one_high[1] & perc_at_least_one_namo <= subprime_mapping$sp_one_high[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_at_least_one_high", "spec" = "6.3_e"),
                                 data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_at_least_two_namo > subprime_mapping$sp_two_low[1] & perc_at_least_two_namo <= subprime_mapping$sp_two_low[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_at_least_two_low", "spec" = "6.3_e"),
                                 data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_at_least_two_namo > subprime_mapping$sp_two_high[1] & perc_at_least_two_namo <= subprime_mapping$sp_two_high[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_at_least_two_high", "spec" = "6.3_e"),
                                 data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_three_namo > subprime_mapping$sp_three_low[1] & perc_three_namo <= subprime_mapping$sp_three_low[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_three_low", "spec" = "6.3_e"),
                                 data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & perc_three_namo > subprime_mapping$sp_three_high[1] & perc_three_namo <= subprime_mapping$sp_three_high[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "sp_perc_three_high", "spec" = "6.3_e"))

openxlsx::write.xlsx(spec_6_3_extensions_maj, file = "results/final/spec_6_3_extensions_maj.xlsx")


### Spec 6.4 (Housing type) ----
spec_6_4_final_maj <- rbind(data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_bottom_tier"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_bottom_tier", "effect" = "both", "stratum" = "all_homes_bottom_tier", "spec" = "6.4"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "all_homes_middle_tier", "spec" = "6.4"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_top_tier"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_top_tier", "effect" = "both", "stratum" = "all_homes_top_tier", "spec" = "6.4"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "single_family_homes"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "single_family_homes", "effect" = "both", "stratum" = "single_family_homes", "spec" = "6.4"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "condo_coop"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "condo_coop", "effect" = "both", "stratum" = "condo_coop", "spec" = "6.4"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "one_bedroom"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "one_bedroom", "effect" = "both", "stratum" = "one_bedroom", "spec" = "6.4"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "two_bedroom"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "two_bedroom", "effect" = "both", "stratum" = "two_bedroom", "spec" = "6.4"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "three_bedroom"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "three_bedroom", "effect" = "both", "stratum" = "three_bedroom", "spec" = "6.4"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "four_bedroom"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "four_bedroom", "effect" = "both", "stratum" = "four_bedroom", "spec" = "6.4"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "five_plus_bedroom"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "five_plus_bedroom", "effect" = "both", "stratum" = "five_plus_bedroom", "spec" = "6.4"))

openxlsx::write.xlsx(spec_6_4_final_maj, file = "results/final/spec_6_4_final_maj.xlsx")


### Spec 6.5 (NRI) ----
load("data/nri.RData")

nri_composite$fips_code <- fips_pad(nri_composite$state_code, nri_composite$county_code)

spec_5_3_panel_maj <- merge(spec_5_3_panel_maj,
                            unique(select(nri_composite, fips_code, risk_score_composite, eal_score_composite, eal_building_value_composite, alr_building_composite)),
                            by = "fips_code", all.x = TRUE)

nri_mapping <- list("score_low" = c(0, 33),
                    "score_medium" = c(33, 67),
                    "score_high" = c(67, 100),
                    "alr_low"= c(0, quantile(spec_5_3_panel_maj$alr_building_composite, 0.33, na.rm = T)),
                    "alr_medium"= c(quantile(spec_5_3_panel_maj$alr_building_composite, 0.33, na.rm = T), quantile(spec_5_3_panel_maj$alr_building_composite, 0.67, na.rm = T)),
                    "alr_high"= c(quantile(spec_5_3_panel_maj$alr_building_composite, 0.67, na.rm = T), quantile(spec_5_3_panel_maj$alr_building_composite, 1, na.rm = T)))

spec_6_5_final_maj <- rbind(data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & risk_score_composite > nri_mapping$score_low[1] & risk_score_composite <= nri_mapping$score_low[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "risk_score_low", "spec" = "6.5"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & risk_score_composite > nri_mapping$score_medium[1] & risk_score_composite <= nri_mapping$score_medium[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "risk_score_medium", "spec" = "6.5"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & risk_score_composite > nri_mapping$score_high[1] & risk_score_composite <= nri_mapping$score_high[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "risk_score_high", "spec" = "6.5"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & eal_score_composite > nri_mapping$score_low[1] & eal_score_composite <= nri_mapping$score_low[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "eal_score_low", "spec" = "6.5"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & eal_score_composite > nri_mapping$score_medium[1] & eal_score_composite <= nri_mapping$score_medium[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "eal_score_medium", "spec" = "6.5"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & eal_score_composite > nri_mapping$score_high[1] & eal_score_composite <= nri_mapping$score_high[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "eal_score_high", "spec" = "6.5"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & alr_building_composite > nri_mapping$alr_low[1] & alr_building_composite <= nri_mapping$alr_low[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "alr_building_low", "spec" = "6.5"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & alr_building_composite > nri_mapping$alr_medium[1] & alr_building_composite <= nri_mapping$alr_medium[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "alr_building_medium", "spec" = "6.5"),
                            data.frame(plm_results(plm(log(zhvi) ~ (nr_dis_lag_0.5_water + nr_dis_lag_0.5_wind + nr_dis_lag_0.5_geological + nr_dis_lag_0.5_extreme_weather + nr_dis_lag_0.5_fire + nr_dis_lag_0.5_human_cause) ^ 2 + unemployment_rate + log(avg_wkly_wage), subset(spec_5_3_panel_maj, data_series == "all_homes_middle_tier" & alr_building_composite > nri_mapping$alr_high[1] & alr_building_composite <= nri_mapping$alr_high[2]), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "stratum" = "alr_building_high", "spec" = "6.5"))

openxlsx::write.xlsx(spec_6_5_final_maj, file = "results/final/spec_6_5_final_maj.xlsx")


### Spec 6.6 (Incident type placebo) ----
load("data/nr_occ_type_panel_maj.RData")

spec_6_6_final_maj <- rbind(data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.5 + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Biological"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Biological", "spec" = "6.6"),
                            data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.25 + nr_dis_lag_0.5_e + nr_dis_lag_1_e + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Biological"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Biological", "spec" = "6.6"),
                            data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.25 + nr_dis_lag_0.5_e + nr_dis_lag_1_e + nr_dis_lag_2_e + nr_dis_lag_3_e + nr_dis_lag_5_e + nr_dis_lag_10_e + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Biological"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Biological", "spec" = "6.6"),
                            data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.5 + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Chemical"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Chemical", "spec" = "6.6"),
                            data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.25 + nr_dis_lag_0.5_e + nr_dis_lag_1_e + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Chemical"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Chemical", "spec" = "6.6"),
                            data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.25 + nr_dis_lag_0.5_e + nr_dis_lag_1_e + nr_dis_lag_2_e + nr_dis_lag_3_e + nr_dis_lag_5_e + nr_dis_lag_10_e + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Chemical"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Chemical", "spec" = "6.6"),
                            data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.5 + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Toxic Substances"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Toxic Substances", "spec" = "6.6"),
                            data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.25 + nr_dis_lag_0.5_e + nr_dis_lag_1_e + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Toxic Substances"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Toxic Substances", "spec" = "6.6"),
                            data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.25 + nr_dis_lag_0.5_e + nr_dis_lag_1_e + nr_dis_lag_2_e + nr_dis_lag_3_e + nr_dis_lag_5_e + nr_dis_lag_10_e + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Toxic Substances"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Toxic Substances", "spec" = "6.6"),
                            data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.5 + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Other"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Other", "spec" = "6.6"),
                            data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.25 + nr_dis_lag_0.5_e + nr_dis_lag_1_e + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Other"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Other", "spec" = "6.6"),
                            data.frame(plm_results(plm(log(zhvi) ~ nr_dis_lag_0.25 + nr_dis_lag_0.5_e + nr_dis_lag_1_e + nr_dis_lag_2_e + nr_dis_lag_3_e + nr_dis_lag_5_e + nr_dis_lag_10_e + unemployment_rate + log(avg_wkly_wage), subset(nr_occ_type_panel_maj, data_series == "all_homes_middle_tier" & incident_type == "Other"), index = c("fips_code", "date"), model = "within", effect = "twoways")), "data_series" = "all_homes_middle_tier", "effect" = "both", "incident_type" = "Other", "spec" = "6.6"))

openxlsx::write.xlsx(spec_6_6_final_maj, file = "results/final/spec_6_6_final_maj.xlsx")
